Spreadsheet preview
Shows what the spreadsheet contains before downloading it. Has a data breakdown on the drive1.txt log on the frequency of message send, the sorted message send, the average time between each message of the same id, DIM behavior when an ID type is completely removed from the log. Useful visualization tool.